Fixup broken code
